rename *.h to *.cpp
